fem que es pugi al server quan el fitxer excedeix una mida determinada

+++ app/src/main/java/epsem/walkinghealth/ServerUploader.java
@@ -52,13 +52,7 @@ public class ServerUploader extends AsyncTask<Void, Void, Void> {
            //HTTP Post - Connexió persistent
            StartConnection();
 
-//           Calendar calendar = Calendar.getInstance();
-//           int hour = calendar.get(Calendar.HOUR_OF_DAY);
-//           int minute = calendar.get(Calendar.MINUTE);
-//
-//           String now = getStringDateTime();
-//           String filename = now +"-"+ hour +"-"+ minute + "_data.txt";// Falta provar i canviar el format del writefile
-           pathToOurFile = new File(Environment.getExternalStorageDirectory(), "WalkingHealth/"+filename);
+           pathToOurFile = new File(Environment.getExternalStorageDirectory(), "WalkingHealth/");//potser peta
            Log.e("app","fitxer: "+pathToOurFile);
            fileInputStream = new FileInputStream(pathToOurFile);
            readFile();
@@ -72,10 +66,10 @@ public class ServerUploader extends AsyncTask<Void, Void, Void> {
            transmitFile();
 
            //Elimina fitxer
-           //if (connection.getResponseCode() ==  OK!){
-           boolean deleted = pathToOurFile.delete();
-           Log.e("App", "fitxer pujat i eliminat: " + deleted);
-           //}
+           if (connection.getResponseCode() ==  200){
+               boolean deleted = pathToOurFile.delete();
+               Log.e("App", "fitxer pujat i eliminat: " + deleted);
+           }
 
        }catch (IOException ioe){
            Log.e("Server", "IOException when connecting"+ioe);
